requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
PyJWT==1.7.1
PySocks==1.6.8
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
six==1.12.0
twilio==6.23.1
urllib3==1.24.1
